requirements.txt
aiohttp==3.9.5
aio-pika==9.4.1
python-dotenv==1.0.1
pydantic==2.7.1